refactor: 볼링 게임 진행 로직 BowlingGame(Controller)에서 분리
